[#31] Refactoring: Added common module.
