+++ auth/build.gradle
@@ -1,4 +1,6 @@
 dependencies {
+    implementation project(':common')
+
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
@@ -6,6 +8,6 @@ dependencies {
     implementation 'org.springframework.cloud:spring-cloud-starter-config'
     implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
     implementation 'org.springframework.cloud:spring-cloud-starter-netflix-ribbon'
-    
+
     runtime 'org.postgresql:postgresql:42.2.5'
 }
