+++ common/src/main/java/common/filter/AuthenticationFilter.java
@@ -1,0 +1,5 @@
+package common.filter;
+
+public class AuthenticationFilter {
+    public static final String DEFAULT_AUTH_COOKIE_NAME = "APP_AUTH";
+}
