+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'd-echo'
 
-include 'config', 'discovery-service', 'echo-service', 'gateway', 'auth'
+include 'config', 'discovery-service', 'echo-service', 'gateway', 'auth', 'common'
